MR fetching: grab refs from origin
This saves some network activity, because both github andgitlab set up
refs in the origin so that you can pull them locally without needing to
get a fork's URL.

This means that the remotes have to have the same names in your config
as they do in your local repository. This whole thing should be a little
more robust about naming of remotes (and like, ensuring they're present
and whatnot).