github: assert that user is member of trusted org
We hired a new person, but neglected to add them to our org on GitHub.
Today, that person went to mint a tag, which didn't correctly include
all the right merge requests!

This happened via a confluence of things:

- If a build step has a `trusted_org_name`, MRs tagged with the right
  label, but by users who are not members of the trusted org, are
  ignored.
- GitHub's API to get organization members only shows public members.
  (Most of Fastmail's GitHub members are private.)
- The person running mint-tag was not, themselves, a member of the
  Fastmail org on GitHub.

This meant that when they went to go mint the tag, mint-tag said "Great,
let's check the trusted members," and only saw the _public_ members of
the org. Many of the tagged MRs were made by private members, so they
simply weren't included in the merge (and since this is the first time
they'd run mint-tag, the log lines saying some MRs were skipped weren't
obvious.)

The solution is straightforward. GitHub provides an API to check
org membership using the provided token's auth, so before we start
merging, we assert that the user running the program is in fact a member
of the trusted org. If not, we bail out immediately, rather than
deploying a subtly wrong tag!

We should do a similar change for GitLab, but GitLab doesn't provide the
same API so it requires slightly more config. (Plus, internal
organization of Fastmail mean we are less likely to be bitten by the
GitLab side of things, so its commit will come later.)