Add support for semilinear merge strategy
This gives the build step a new config option, use_semilinear_merge. If
that's set, for every MR, we'll rebase it and create a merge commit as
we go. This is different than the current `rebase` option, which for N
MRs to be merged into main, rebases all N on main and then does a single
octopus merge; semilinear merge will create N merge commits.

There's a minor bit of refactoring here, because I realized we can just
set the author/committer env vars in one place rather than doing it
in multiple places. (Setting GIT_AUTHOR_EMAIL is safe to do because git
will not rewrite commit authors unless you ask it to, which we do not.)