Local conf: support more auto-configuration
Before, your local conf had to contain `upstream_base`, which was
something like `github/main`. Now instead, you can have
`upstream_remote` and `use_upstream_default_branch`, so you can say
`upstream_remote = "github"`, and we'll fetch whatever the default
branch name is on the remote (likely something like `main` or `master`).

This commit also allows you to specify a `base_dir` rather than a `path`
in your local config. If you do so, we'll use the path to the repo
underneath your base dir as our working directory. That is: if you're
working on `fastmail/mint-tag`, you can set `base_dir = "/tmp/git-work"`
and we'd clone and do all of our work in
`/tmp/git-work/fastmail/mint-tag`. (This is in service of something
we'll add in the next commit.)