Fix behavior of force-push for forks
The change introduced in 3a339676 does not work for the normal
GitHub/GitLab case, where changes are introduced in a fork and merged
into the golden repository. This commit fixes that by adding a new
property on merge request objects, `force_push_url`, and populating it
appropriately.

Imagine we're working on the repo `fastmail/mint-tag`, and I introduce a
change on `michael/mint-tag`, in MR 42. Before, we'd do this:

- http fetch MR 42
- git fetch it, using the magic ref generated for merge requests
- do our rebase and merge
- attempt to force push to fastmail/my-feature-branch

This last step was failing, because my-feature-branch never existed on
fastmail/mint-tag; it only exists on michael/mint-tag! Now, we'll
correctly push to michael/mint-tag, using the ssh URL we populate during
the http fetch.