Wait for remote to notice when pushing rebased MRs
When you force-push to an MR (in GitLab, at least), there is often a lag
between the time you do so (with git) and the the time GitLab takes to
notice it (via http). That means that if you rebase an MR, force-push to
the branch, then immediately merge it and push to the golden repo,
GitLab *won't notice* and thus won't mark the MR as merged, but instead
it stays open and says "no changes," which sucks.

This commit also sucks, which is just "sleep in a loop until the remote
is up to date".