MergeRequest: pass --force-rebase when rebasing
We have seen behavior in production where git seems to do the _wrong
thing_ when rebasing a branch that is already up to date. Here's a
sequence of events.

1. git checkout main (main is now at 0f584dd5, say)
2. we have some other branch, topic (98d691943)
3. git merge-base main 98d691943  (returns 0f584dd5, which is main)
4. git rebase 0f584dd5 98d691943 (reports: 'Current branch 98d691943 is up to date.')
5. git status reports we're on branch main
6. ...but main has _changed_, and is now 98d691943

Step four here legitimately seems like a bug in git. According to the
docs, `git rebase main topic` is simply shorthand for
`git checkout topic; git rebase main`, and at the end, you're meant to
be on branch topic. (Or, in the usual case where we're interacting with
shas directly, in a detached HEAD on the newly rebased branch). This
_is_ the case when topic is not up to date. What git does when it _is_
up to date seems to be to fast-forward main to the new commit, leaving
you on main!

The upshot of this is that after that, when we went to merge
semi-linearly, we'd say `git merge --no-ff 98d691943`, and git would say
"already up-to-date", because we'd already fast-forwarded the main
branch! This is, if not a bug, extremely surprising to me (at the very
least). But we can avoid the problem entirely by passing
`--force-rebase` to the rebase invocation, which will refuse to
fast-forward and always rewrite commits.