Add setting to delete merged branches
GitLab has a checkbox on MRs to delete branches when merged, but it
doesn't work if you merge the branch in git and not by clicking the
merge button in the UI. (Sigh, GitLab.) So now, we'll check for that
checkbox on the MRs and delete the remote branches after we merge them.

GitHub does this correctly, of course, and the setting actually deletes
the branches when they're merged in any way at all.