Check for truthiness of push_spec before pushing
This is just a safety because I nearly shot my foot off while testing
something locally. If there's no push_spec defined in config, and you
have a release-mode config, we generate one for you that pushes back to
the canonical remote. That's very useful, except that I was testing a
mint-tag config tweak and inadvertently pushed my trial merge to
upstream, which I did not want to do!

I thought this was safe, because I had commented out the push_spec in my
config, thinking that was sufficient to prevent the final push. (For
non-release-mode configs, it is.) This commit makes this safer by also
checking for its truthiness. That is, if you don't provide a push spec
in a release-mode config, we'll generate one for you; if you _really_
don't want to push, you can provide `push_spec = false` and we won't
push at all.