setup: abort rebases in non-interactive mode
In production, we hit a problem where a mint with explicit merge
requests failed, because the two merge requests conflicted with one
another. Our config has rebase=true, and so the failure happened after
we'd merged the first and were rebasing the second. That means when we
exited, we were mid-rebase, and so the next attempt (after fixing the
merge conflict) failed with

    fatal: It seems that there is already a rebase-merge directory, and
    I wonder if you are in the middle of another rebase.  If that is the
    case, please try
      git rebase (--continue | --abort | --skip)
    If that is not the case, please
      rm -fr ".git/rebase-merge"
    and run me again.  I am stopping in case you still have something
    valuable there.

Indeed, there was! Let's detect this case and just remove the
rebase-merge directory if it exists. But only do it if running
non-interactively, because it's potentially dangerous.