PLAT-1751: Tag pruning
This allows mint-tag to prune old tags. We have too many tags and we
don't want to add to the fire.

Specifically, this introduces a new `cleanup_tag_days` config option
that can be passed for build steps that have `push_tag_to` set and if
set to a nonzero number all tags that match the `tag_prefix` and have a
commit date older than (roughly) that number of days will be deleted
from the remote.